refactor: cleanup, rename and move files and names Move components into components folder and utilities into utilitsfolder. Rename long and hard to read names. Prioritize non relative path over relative paths.
