+++ src/components/PanelOptions/ImportExport.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { StandardEditorProps } from '@grafana/data';
-import { OptionsInterface, EditorLanguageType } from './types';
-import { TextPanelEditor } from './TextPanelEditor';
+import { OptionsInterface, EditorLanguageType } from 'types';
+import { TextEditor } from 'components/TextEditor';
 import { Button, Input } from '@grafana/ui';
 import { exportFile, contentType } from 'utils/exportFile';
 import { readFile } from 'utils/readFile';
@@ -34,7 +34,7 @@ const Spacer: React.FC = () => {
   return <div style={{ marginBottom: '6px' }}></div>;
 };
 
-export const PanelOptionImportedPanelOptions: React.FC<Props> = ({ value, item, onChange, context }) => {
+export const ImportExportOption: React.FC<Props> = ({ value, item, onChange, context }) => {
   if (context.options) {
     const options = { ...context.options };
     delete options.importedPanelOptions;
@@ -52,7 +52,7 @@ export const PanelOptionImportedPanelOptions: React.FC<Props> = ({ value, item, 
     <div>
       <Input css={{}} type="file" onChange={(e) => importPanelOptions(e.currentTarget.files, updatePanelOptions)} />
       <Spacer />
-      <TextPanelEditor language={item.settings?.language} value={value} onChange={updatePanelOptions} />
+      <TextEditor language={item.settings?.language} value={value} onChange={updatePanelOptions} />
       <Spacer />
       <Button onClick={() => downloadJsonFile(value)}>Download as JSON file</Button>
     </div>
