+++ src/components/PanelOptions/TextEditor.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { StandardEditorProps } from '@grafana/data';
-import { TextPanelEditor } from 'TextPanelEditor';
+import { TextEditor } from 'components/TextEditor';
 import { EditorLanguageType, EditorCodeType } from 'types';
 
 interface Settings {
@@ -9,6 +9,6 @@ interface Settings {
 
 interface Props extends StandardEditorProps<EditorCodeType, Settings> {}
 
-export const PanelOptionCode: React.FC<Props> = ({ value, item, onChange }) => {
-  return <TextPanelEditor language={item.settings?.language} value={value} onChange={(code) => onChange(code)} />;
+export const TextEditorOption: React.FC<Props> = ({ value, item, onChange }) => {
+  return <TextEditor language={item.settings?.language} value={value} onChange={(code) => onChange(code)} />;
 };
