+++ src/components/TextEditor/declarations/codeData.tsx
