+++ src/components/TextEditor/declarations/customProperties.tsx
