+++ src/components/TextEditor/declarations/data.tsx
