+++ src/components/TextEditor/declarations/dataFrame.tsx
