+++ src/components/TextEditor/declarations/dataQuery.tsx
