+++ src/components/TextEditor/declarations/dataQueryRequest.tsx
