+++ src/components/TextEditor/declarations/fieldColor.tsx
