+++ src/components/TextEditor/declarations/getLocationSrv.tsx
