+++ src/components/TextEditor/declarations/getTemplateSrv.tsx
