+++ src/components/TextEditor/declarations/index.tsx
