+++ src/components/TextEditor/declarations/options.tsx
