+++ src/components/TextEditor/declarations/scopedVars.tsx
