+++ src/components/TextEditor/declarations/textEditorDeclarations.test.tsx
@@ -1,4 +1,4 @@
-import textEditorDeclarations from './';
+import textEditorDeclarations from '.';
 
 describe('Text editor declarations', () => {
   it('is a valid string', () => {
