+++ src/components/TextEditor/declarations/theme.tsx
