+++ src/components/TextEditor/declarations/thresholds.tsx
