+++ src/components/TextEditor/declarations/timeRanges.tsx
