+++ src/components/TextEditor/declarations/valueMapping.tsx
