+++ src/fonts.scss
@@ -1,6 +1,6 @@
 @font-face {
   font-family: 'Open Sans';
-  src: url('./fonts/OpenSans-Light.woff2') format('woff2');
+  src: url('fonts/OpenSans-Light.woff2') format('woff2');
   font-weight: 300;
   font-style: normal;
   font-display: swap;
@@ -8,7 +8,7 @@
 
 @font-face {
   font-family: 'Open Sans';
-  src: url('./fonts/OpenSans-Regular.woff2') format('woff2');
+  src: url('fonts/OpenSans-Regular.woff2') format('woff2');
   font-weight: normal;
   font-style: normal;
   font-display: swap;
@@ -16,7 +16,7 @@
 
 @font-face {
   font-family: 'Open Sans';
-  src: url('./fonts/OpenSans-SemiBold.woff2') format('woff2');
+  src: url('fonts/OpenSans-SemiBold.woff2') format('woff2');
   font-weight: 600;
   font-style: normal;
   font-display: swap;
@@ -24,7 +24,7 @@
 
 @font-face {
   font-family: 'Open Sans';
-  src: url('./fonts/OpenSans-Bold.woff2') format('woff2');
+  src: url('fonts/OpenSans-Bold.woff2') format('woff2');
   font-weight: bold;
   font-style: normal;
   font-display: swap;
